Move schema generation into @pikku/inspector with virtual file approach
Schema generation (JSON Schema from TS types + Zod schemas) now lives in
@pikku/inspector/schema-generator, making InspectorState the single source
of truth with a new `schemas` field. Uses ts-json-schema-generator's
tsProgram option with a virtual CompilerHost to serve custom types
in-memory — no temp file writes needed.

CLI commands (pikkuSchemas, pikkuOpenAPI) are now pure file writers that
read from state.schemas. Contract hash computation stays in the CLI.
