+++ packages/inspector/src/schema-generator.ts
@@ -1,0 +1,1 @@
+export { generateAllSchemas } from './utils/schema-generator.js'
